Added Connect-DSClientSession Cmdlet
Additional Base Class for Session Handling Cmdlets
